Add files via upload
This was a test function used to run Nodefile, if one wants to test varying outputs, please modify the file to take in the specific array of ones choosing. In addition, if one is not working with dates, it is recommended to change line 73 on Nodefile.js, if not comment it out completely.